- port mathmore changes of dev brunch done by David
- improve tests to correctly return a status 
- some clean up of minimizer classes (use directly ROOT::Math::IMultiGenFunction and ROOT::Math::IMultiGradFunction instead of typedef's)


git-svn-id: http://root.cern.ch/svn/root/trunk@23619 27541ba8-7e3a-0410-8455-c3a389f83636

+++ math/mathmore/test/testRandomDist.cxx
@@ -247,7 +247,7 @@ void testPoissonCLHEP( R & r, double mu,TH1D & h) {
   //  r.SetSeed(0);
   for (int i = 0; i < n; ++i) {
     //int n = RandPoisson::shoot(mu + RandFlat::shoot());
-    int n = r(mu);
+    int n = static_cast<unsigned int> ( r(mu) ) ;
     if (fillHist)
       h.Fill( double(n) );
   }
